pyproject.toml
[tool.poetry]
name = "aod-project"
version = "1.0.0"
description = "CDK project for creating resources and components"
authors = [
    "Jose Nunes <nunesjos@amazon.lu>",
]
readme = "README.md"
package-mode=false

[tool.bandit]
exclude_dirs = [".venv", "cdk/cdk.out" ]
verbose = true
skips = ["B101"]

[tool.black]
line-length = 100
preview = true
skip-magic-trailing-comma = true


[tool.isort]
profile = "black"
line_length = 100

[tool.pylint]
# UPDATE: Aligned with flake8 and black
max-line-length = 100
fail-under = 8
ignore-paths = "cdk/cdk.out/*"

[tool.pylint.messages_control]
disable = [
    "E0401",
    "R0801",
    "R0903",
    "R0913",
    "R0914",
    "W0602",
    "W0603"
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
aws-cdk-lib = "2.130.0"
boto3 = "1.34.51"
cdk-nag = "2.28.48"
constructs = "10.3.0"
pre-commit = "3.6.2"
fpdf = "^1.7.2"
tox = "^4.23.2"
pygit2 = "^1.16.0"

[tool.poetry.group.security.dependencies]
pip-audit = "2.7.1"

[tool.poetry.group.lint.dependencies]
flake8 = "7.0.0"
black = "24.2.0"
pylint = "3.1.0"

[tool.poetry.group.test.dependencies]
coverage = "7.4.3"
pytest = "8.0.2"
pytest-cov = "4.1.0"
assertpy = "1.1"
pytest-env = "^1.1.3"

[tool.yamlfix]
line_length = 240
sequence_style = "keep_style"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
